requirements.txt
eval_metrics==0.0.1
hydra-core==1.3.2
lightning==2.6.0
numpy==2.3.5
omegaconf==2.3.0
pandas==2.3.3
Pillow==12.0.0
pytorch_lightning==2.6.0
scikit_learn==1.8.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
